Moved files to components/list folder
