+++ app/components/lists/ListItemDeleteAction.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { StyleSheet, View } from 'react-native';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 
-import colors from '../config/colors';
+import colors from '../../config/colors';
 import { TouchableWithoutFeedback } from 'react-native-gesture-handler';
