+++ app/screens/ListingDetailsScreen.js
@@ -4,7 +4,7 @@ import { View, StyleSheet, Image } from 'react-native'
 
 
 import AppText from '../components/AppText';
-import ListItem from '../components/ListItem';
+import ListItem from '../components/lists/ListItem';
 import colors from '../config/colors';
 
 function ListingDetailsScreen(props) {
